# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserProfile": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"aboutMe": {
						"type": "string"
					},
					"phoneCountryCode": {
						"type": "number",
						"format": "double"
					},
					"phoneNumber": {
						"type": "number",
						"format": "double"
					},
					"yearOfBirth": {
						"type": "number",
						"format": "double"
					},
					"location": {
						"type": "string"
					},
					"userType": {
						"type": "string"
					},
					"themeColor": {
						"type": "string"
					},
					"spotifyPlaylistUrl": {
						"type": "string"
					},
					"onboardingCompleted": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"firstName",
					"lastName",
					"location",
					"userType",
					"onboardingCompleted",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProfileDto": {
				"description": "DTO for updating user profile\r\n\r\nUsed to update user profile information. All fields are optional - only\r\nprovided fields will be updated. The user_id is automatically extracted\r\nfrom the authentication token.",
				"properties": {
					"firstName": {
						"type": "string",
						"description": "First name (1-50 characters)",
						"example": "John"
					},
					"lastName": {
						"type": "string",
						"description": "Last name (1-50 characters)",
						"example": "Doe"
					},
					"bio": {
						"type": "string",
						"description": "Short bio (max 100 characters)",
						"example": "Musician from New York"
					},
					"aboutMe": {
						"type": "string",
						"description": "Detailed about me section (max 500 characters)",
						"example": "I'm a professional musician with 10 years of experience..."
					},
					"avatarUrl": {
						"type": "string",
						"description": "URL to user's avatar image",
						"example": "https://example.com/avatar.jpg"
					},
					"location": {
						"type": "string",
						"description": "User's location (1-100 characters)",
						"example": "New York, NY"
					},
					"themeColor": {
						"type": "string",
						"description": "Theme color in hex format",
						"example": "#FF5733"
					},
					"spotifyPlaylistUrl": {
						"type": "string",
						"description": "URL to Spotify playlist",
						"example": "https://open.spotify.com/playlist/..."
					},
					"phoneCountryCode": {
						"type": "number",
						"format": "double",
						"description": "Phone country code (1-999)",
						"example": 1
					},
					"phoneNumber": {
						"type": "number",
						"format": "double",
						"description": "Phone number (4-15 digits)",
						"example": 1234567890
					},
					"onboardingCompleted": {
						"type": "boolean",
						"description": "Whether user has completed onboarding",
						"example": true
					}
				},
				"type": "object",
				"additionalProperties": false,
				"example": {
					"firstName": "John",
					"lastName": "Doe",
					"bio": "Musician from New York",
					"location": "New York, NY",
					"themeColor": "#FF5733"
				}
			},
...
			"NotificationResponse": {
				"description": "API response format for a notification\r\nRepresents a notification with actor information",
				"properties": {
					"id": {
						"type": "string"
					},
					"recipientId": {
						"type": "string"
					},
					"actorId": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string"
					},
					"entityType": {
						"type": "string",
						"nullable": true
					},
					"entityId": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string",
						"nullable": true
					},
					"actionUrl": {
						"type": "string",
						"nullable": true
					},
					"isRead": {
						"type": "boolean",
						"nullable": true
					},
					"isArchived": {
						"type": "boolean",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"nullable": true
					},
					"readAt": {
						"type": "string",
						"nullable": true
					},
					"sentViaWebsocket": {
						"type": "boolean",
						"nullable": true
					},
					"actor": {
						"properties": {
							"avatarUrl": {
								"type": "string",
								"nullable": true
							},
							"lastName": {
								"type": "string",
								"nullable": true
							},
							"firstName": {
								"type": "string",
								"nullable": true
							},
							"username": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"username",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"recipientId",
					"type",
					"title",
					"isRead",
					"isArchived",
					"createdAt",
					"sentViaWebsocket"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateNotificationDto": {
				"description": "DTO for updating a notification\r\n\r\nUsed when marking a notification as read or unread.",
				"properties": {
					"isRead": {
						"type": "boolean",
						"description": "Whether the notification is read",
						"example": true
					}
				},
				"required": [
					"isRead"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"isRead": true
				}
			},
...
			"CommentResponse": {
				"description": "API response format for a comment\r\nRepresents a comment on a post with author information",
				"properties": {
					"id": {
						"type": "string"
					},
					"postId": {
						"type": "string"
					},
					"authorId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"nullable": true
					},
					"updatedAt": {
						"type": "string",
						"nullable": true
					},
					"author": {
						"properties": {
							"avatarUrl": {
								"type": "string",
								"nullable": true
							},
							"lastName": {
								"type": "string",
								"nullable": true
							},
							"firstName": {
								"type": "string",
								"nullable": true
							},
							"username": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"username",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"postId",
					"authorId",
					"content",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "LineUp API",
		"version": "1.0.0",
		"description": "API documentation for the LineUp application",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Pinnaacle"
		}
	},
	"paths": {
		"/users/{username}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "User profile (public or private based on authentication)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfile"
								}
							}
						}
					}
				},
				"description": "Get user profile by username\r\n\r\nRetrieves a user's profile information. Returns public profile data by default.\r\nIf authenticated and viewing your own profile, returns additional private fields\r\nsuch as phone number, year of birth, etc.\r\n\r\nAuthentication is optional - endpoint can be accessed without a token for public profiles.",
				"summary": "Get user profile",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The username of the user to retrieve",
						"in": "path",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Updated user profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfile"
								}
							}
						}
					}
				},
				"description": "Update own profile\r\n\r\nUpdates the authenticated user's profile information. Only the profile owner\r\ncan update their own profile. All fields in the request body are optional -\r\nonly provided fields will be updated.",
				"summary": "Update user profile",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The username of the profile to update (must match authenticated user)",
						"in": "path",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The profile fields to update (all optional)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileDto",
								"description": "The profile fields to update (all optional)"
							}
						}
					}
				}
			}
		},
...
		"/notifications": {
			"get": {
				"operationId": "GetNotifications",
				"responses": {
					"200": {
						"description": "Notifications with pagination cursor",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"nextCursor": {
											"type": "string"
										},
										"notifications": {
											"items": {
												"$ref": "#/components/schemas/NotificationResponse"
											},
											"type": "array"
										}
									},
									"required": [
										"notifications"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get notifications for the authenticated user\r\n\r\nReturns notifications for the authenticated user with optional filters.\r\nSupports filtering by type and unread status, with cursor-based pagination.\r\nArchived notifications are excluded by default.",
				"summary": "Get user notifications",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Filter by notification type",
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If true, only return unread notifications",
						"in": "query",
						"name": "unreadOnly",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Cursor for pagination (ISO timestamp)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Maximum number of notifications to return (1-100, default: 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/notifications/{notificationId}": {
			"get": {
				"operationId": "GetNotification",
				"responses": {
					"200": {
						"description": "The notification with actor information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationResponse"
								}
							}
						}
					}
				},
				"description": "Get a notification by ID\r\n\r\nReturns a specific notification with actor information.\r\nOnly the recipient can access their own notifications.",
				"summary": "Get notification by ID",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The UUID of the notification to retrieve",
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateNotification",
				"responses": {
					"200": {
						"description": "The updated notification",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationResponse"
								}
							}
						}
					}
				},
				"description": "Update a notification\r\n\r\nMarks a notification as read or unread.\r\nOnly the recipient can update their own notifications.",
				"summary": "Mark notification as read/unread",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The UUID of the notification to update",
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Update data including isRead status",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateNotificationDto",
								"description": "Update data including isRead status"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteNotification",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete a notification\r\n\r\nPermanently deletes a notification.\r\nOnly the recipient can delete their own notifications.",
				"summary": "Delete a notification",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The UUID of the notification to delete",
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/comments/user/{userId}": {
			"get": {
				"operationId": "GetUserComments",
				"responses": {
					"200": {
						"description": "Array of comments with author details",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CommentResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all comments by a user\r\n\r\nReturns all comments created by a specific user, ordered by creation date (newest first).\r\nEach comment includes the author's profile information.",
				"summary": "Get comments by user",
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [
					{
						"description": "The UUID of the user to retrieve comments for",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}